requirements.txt
opencv-python
tensorflow
numpy
pillow